Reorganize shared utilities folder
- move save_output to shared folder and name save_geo_output
- split up gdf utils and status manager into separate files
