+++ src/bikespace_data/utilities/__init__.py
@@ -1,3 +1,4 @@
 """Shared utility functions used by more than one dataset."""
 
-from .utilities import StatusManager, status_manager_date_format, dt_cols_to_str
+from .status_manager import StatusManager, status_manager_date_format
+from .gdf_utils import dt_cols_to_str, save_geo_output
